backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5
